requirements.txt
Django==4.2.1
PyJWT==2.7.0
